Review: applied suggestion to move self_test stuff from the package module to the probe one
