+++ pkg/security/module/module.go
@@ -68,7 +68,7 @@ type Module struct {
 	rulesLoaded      func(rs *rules.RuleSet, err *multierror.Error)
 	policiesVersions []string
 
-	selfTester *SelfTester
+	selfTester *sprobe.SelfTester
 	reloader   *debouncer.Debouncer
 }
 
@@ -212,23 +212,6 @@ func (m *Module) getEventTypeEnabled() map[eval.EventType]bool {
 	return enabled
 }
 
-func logMultiErrors(msg string, m *multierror.Error) {
-	var errorLevel bool
-	for _, err := range m.Errors {
-		if rErr, ok := err.(*rules.ErrRuleLoad); ok {
-			if !errors.Is(rErr.Err, rules.ErrEventTypeNotEnabled) {
-				errorLevel = true
-			}
-		}
-	}
-
-	if errorLevel {
-		log.Errorf(msg, m.Error())
-	} else {
-		log.Warnf(msg, m.Error())
-	}
-}
-
 func getPoliciesVersions(rs *rules.RuleSet) []string {
 	var versions []string
 
@@ -305,9 +288,9 @@ func (m *Module) Reload() error {
 	loadErr := rules.LoadPolicies(policiesDir, ruleSet)
 
 	if loadErr.ErrorOrNil() != nil {
-		logMultiErrors("error while loading policies: %+v", loadErr)
+		seclog.LogRuleLoadingErrors("error while loading policies: %+v", loadErr)
 	} else if loadApproversErr.ErrorOrNil() != nil {
-		logMultiErrors("error while loading policies for Approvers: %+v", loadApproversErr)
+		seclog.LogRuleLoadingErrors("error while loading policies for Approvers: %+v", loadApproversErr)
 	}
 
 	monitor := m.probe.GetMonitor()
@@ -428,7 +411,7 @@ func (m *Module) RuleMatch(rule *rules.Rule, event eval.Event) {
 		return append(tags, m.probe.GetResolvers().TagsResolver.Resolve(id)...)
 	}
 
-	if !m.selfTester.isExpectedEvent(rule, event) {
+	if !m.selfTester.IsExpectedEvent(rule, event) {
 		m.SendEvent(rule, event, extTagsCb, service)
 	}
 }
@@ -548,7 +531,7 @@ func NewModule(cfg *sconfig.Config, opts ...Opts) (module.Module, error) {
 	// custom limiters
 	limits := make(map[rules.RuleID]Limit)
 
-	var selfTester = NewSelfTester()
+	var selfTester = sprobe.NewSelfTester()
 
 	m := &Module{
 		config:       cfg,
